Code cleanup - compilation warnings

+++ src/protocol/multichaintx.cpp
@@ -659,7 +659,7 @@ bool AcceptAssetGenesisFromPredefinedIssuers(const CTransaction &tx,
             {
                 memcpy(issuer_buf,issuers[i].begin(),sizeof(uint160));
                 mc_PutLE(issuer_buf+sizeof(uint160),&issuer_flags[i],4);
-                if(i < mc_gState->m_Assets->MaxStoredIssuers())
+                if((int)i < mc_gState->m_Assets->MaxStoredIssuers())
                 {
                     mc_gState->m_TmpScript->SetSpecialParamValue(MC_ENT_SPRM_ISSUER,issuer_buf,sizeof(issuer_buf));            
                 }
@@ -1799,7 +1799,7 @@ bool AcceptMultiChainTransaction(const CTransaction& tx,
                 {
                     memcpy(opener_buf,openers[i].begin(),sizeof(uint160));
                     mc_PutLE(opener_buf+sizeof(uint160),&opener_flags[i],4);
-                    if(i < mc_gState->m_Assets->MaxStoredIssuers())
+                    if((int)i < mc_gState->m_Assets->MaxStoredIssuers())
                     {
                         mc_gState->m_TmpScript->SetSpecialParamValue(MC_ENT_SPRM_ISSUER,opener_buf,sizeof(opener_buf));            
                     }
@@ -2636,7 +2636,7 @@ bool AcceptAssetGenesis(const CTransaction &tx,int offset,bool accept,string& re
             if(stored_issuers.count(issuers[i]) == 0)
             {
                 memcpy(issuer_buf,issuers[i].begin(),sizeof(uint160));
-                if(i < mc_gState->m_Assets->MaxStoredIssuers())
+                if((int)i < mc_gState->m_Assets->MaxStoredIssuers())
                 {
                     mc_gState->m_TmpScript->SetSpecialParamValue(MC_ENT_SPRM_ISSUER,issuer_buf,sizeof(issuer_buf));            
                 }
